+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
  - Fixed bug where organization dropdown could be shown to a user with no organizations
  - Fixed a bug where users can invite the same user to their organization twice.
+ - Fixed a bug where users could accidentally create duplicate organizations
 ### Changed
  - Existing users can now be invited to new and existing organizations.